Fixed weird dependencies between project and config, project has a config now
