+++ builder/project.py
@@ -16,8 +16,10 @@ import glob
 import os
 import sys
 
-from config import produce_config
+from data import *
+from host import current_platform
 from scripts import Scripts
+from util import replace_variables
 
 
 def looks_like_code(path):
@@ -29,6 +31,124 @@ def looks_like_code(path):
     if readme:
         return True
     return False
+
+
+def _apply_value(obj, key, new_value):
+    """ Merge values according to type """
+    key_type = type(new_value)
+    if key_type == list:
+        # Apply the config's value before the existing list
+        obj[key] = new_value + obj[key]
+
+    elif key_type == dict:
+        # Iterate each element and recursively apply the values
+        for k, v in new_value.items():
+            _apply_value(obj[key], k, v)
+
+    else:
+        # Unsupported type, just use it
+        obj[key] = new_value
+
+
+def produce_config(build_spec, project, **additional_variables):
+    """ Traverse the configurations to produce one for the given spec """
+    platform = current_platform()
+
+    defaults = {
+        'hosts': HOSTS,
+        'targets': TARGETS,
+        'compilers': COMPILERS,
+    }
+
+    # Build the list of config options to poll
+    configs = []
+
+    # Processes a config object (could come from a file), searching for keys hosts, targets, and compilers
+    def process_config(config):
+
+        def process_element(map, element_name, instance):
+            if not map:
+                return
+
+            element = map.get(element_name)
+            if not element:
+                return
+
+            new_config = element.get(instance)
+            if not new_config:
+                return
+
+            configs.append(new_config)
+
+            config_archs = new_config.get('architectures')
+            if config_archs:
+                config_arch = config_archs.get(build_spec.arch)
+                if config_arch:
+                    configs.append(config_arch)
+
+            return new_config
+
+        # Get defaults from platform (linux) then override with host (al2, manylinux, etc)
+        if platform != build_spec.host:
+            process_element(config, 'hosts', platform)
+        process_element(config, 'hosts', build_spec.host)
+        process_element(config, 'targets', build_spec.target)
+
+        compiler = process_element(config, 'compilers', build_spec.compiler)
+        process_element(compiler, 'versions', build_spec.compiler_version)
+
+    # Process defaults first
+    process_config(defaults)
+
+    # then override with config file
+    project_config = project.config
+    process_config(project_config)
+    if project_config not in configs:
+        configs.append(project_config)
+
+    new_version = {
+        'spec': build_spec,
+    }
+    # Iterate all keys and apply them
+    for key, default in KEYS.items():
+        new_version[key] = default
+
+        for config in configs:
+            override_key = '!' + key
+            if override_key in config:
+                # Handle overrides
+                new_version[key] = config[override_key]
+
+            elif key in config:
+                # By default, merge all values (except strings)
+                _apply_value(new_version, key, config[key])
+
+    # Default variables
+    replacements = {
+        'host': build_spec.host,
+        'compiler': build_spec.compiler,
+        'version': build_spec.compiler_version,
+        'target': build_spec.target,
+        'arch': build_spec.arch,
+        'cwd': os.getcwd(),
+        **additional_variables,
+    }
+
+    # Pull variables from the configs
+    for config in configs:
+        if 'variables' in config:
+            variables = config['variables']
+            assert type(variables) == dict
+
+            # Copy into the variables list
+            for k, v in variables.items():
+                replacements[k] = v
+
+    # Post process
+    new_version = replace_variables(new_version, replacements)
+    new_version['variables'] = replacements
+
+    return new_version
 
 
 class Project(object):
@@ -85,7 +205,7 @@ class Project(object):
                 consumers.append(c)
         return consumers
 
-    def get_config(spec, **additional_vars):
+    def get_config(self, spec, **additional_vars):
         return produce_config(spec, self, ** additional_vars)
 
     # project cache
